+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "Myriad"
+name = "myriad"
 version = "0.1.0"
 edition = "2021"
