+++ src/creature/mod.rs
@@ -1,0 +1,2 @@
+pub mod bodypart;
+pub mod creature;
